[PM-1880] Added ImportOrganizationCommand
