+++ bitwarden_license/src/Scim/Users/PostUserCommand.cs
@@ -3,6 +3,7 @@ using Bit.Core.Enums;
 using Bit.Core.Exceptions;
 using Bit.Core.Models.Data;
 using Bit.Core.Models.Data.Organizations.OrganizationUsers;
+using Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Bit.Core.Utilities;
@@ -17,15 +18,18 @@ public class PostUserCommand : IPostUserCommand
     private readonly IOrganizationUserRepository _organizationUserRepository;
     private readonly IOrganizationService _organizationService;
     private readonly IScimContext _scimContext;
+    private readonly IInviteOrganizationUserCommand _inviteOrganizationUserCommand;
 
     public PostUserCommand(
         IOrganizationUserRepository organizationUserRepository,
         IOrganizationService organizationService,
-        IScimContext scimContext)
+        IScimContext scimContext,
+        IInviteOrganizationUserCommand inviteOrganizationUserCommand)
     {
         _organizationUserRepository = organizationUserRepository;
         _organizationService = organizationService;
         _scimContext = scimContext;
+        _inviteOrganizationUserCommand = inviteOrganizationUserCommand;
     }
 
     public async Task<OrganizationUserUserDetails> PostUserAsync(Guid organizationId, ScimUserRequestModel model)
@@ -80,7 +84,7 @@ public class PostUserCommand : IPostUserCommand
             throw new ConflictException();
         }
 
-        var invitedOrgUser = await _organizationService.InviteUserAsync(organizationId, EventSystemUser.SCIM, email,
+        var invitedOrgUser = await _inviteOrganizationUserCommand.InviteUserAsync(organizationId, EventSystemUser.SCIM, email,
             OrganizationUserType.User, false, externalId, new List<CollectionAccessSelection>(), new List<Guid>());
         var orgUser = await _organizationUserRepository.GetDetailsByIdAsync(invitedOrgUser.Id);
