+++ src/Api/Public/Controllers/OrganizationController.cs
@@ -3,6 +3,7 @@ using Bit.Api.Models.Public.Request;
 using Bit.Api.Models.Public.Response;
 using Bit.Core.Context;
 using Bit.Core.Exceptions;
+using Bit.Core.OrganizationFeatures.Import.Interfaces;
 using Bit.Core.Services;
 using Bit.Core.Settings;
 using Microsoft.AspNetCore.Authorization;
@@ -17,15 +18,18 @@ public class OrganizationController : Controller
     private readonly IOrganizationService _organizationService;
     private readonly ICurrentContext _currentContext;
     private readonly GlobalSettings _globalSettings;
+    private readonly IImportOrganizationCommand _importOrganizationCommand;
 
     public OrganizationController(
         IOrganizationService organizationService,
         ICurrentContext currentContext,
-        GlobalSettings globalSettings)
+        GlobalSettings globalSettings,
+        IImportOrganizationCommand importOrganizationCommand)
     {
         _organizationService = organizationService;
         _currentContext = currentContext;
         _globalSettings = globalSettings;
+        _importOrganizationCommand = importOrganizationCommand;
     }
 
     /// <summary>
@@ -46,7 +50,7 @@ public class OrganizationController : Controller
             throw new BadRequestException("You cannot import this much data at once.");
         }
 
-        await _organizationService.ImportAsync(
+        await _importOrganizationCommand.ImportAsync(
             _currentContext.OrganizationId.Value,
             null,
             model.Groups.Select(g => g.ToImportedGroup(_currentContext.OrganizationId.Value)),
